Correct defect in Dummy auth
Was setting the user ID during the loginStart, but in real providers
it's not done until the end (once the external call returns and provides
an identifier that can be correlated with a user account).  The issue
was causing meant a login failure in DummyAuth was still switching the
PHP.Gt userID even though HybridAuth never accepted that the user was
authenticated.