Auto-set the sass debug-info flag if not in Production
Allows sass to write-out the debug info to the processed CSS files -
info that can then be consumed by the firesass plugin etc to debug the
sass/ css files.

Rather than adding another setting, this commit sets the flag based on
the existing isProduction flag - so if isProduction is set to false, the
debug info is compiled-in/

Fixes #147.